move ai-grading.ts back outside
